Set up Visual Studio project files.
A VS Solution has been created as well as a Project file for the engine.

Additionally, .gitignore and .gitattributes files have been added as well as a basic CMakeLists.txt (there is only a Solution global CMakeLists.txt file but not for the NostraEngine project).

Also, the configuration is not complete yet, as there are no 3rd party libraries set up yet (and neither is NostraUtils).